+++ epstein_files/documents/config/categories/girls.py
@@ -1,6 +1,8 @@
 from epstein_files.documents.documents.categories import Interesting, Neutral, Uninteresting
+from epstein_files.documents.config.communication_cfg import CommunicationCfg
 from epstein_files.documents.config.config_builder import letter
-from epstein_files.documents.config.doc_cfg import NO_TRUNCATE, CommunicationCfg, DocCfg, EmailCfg
+from epstein_files.documents.config.doc_cfg import NO_TRUNCATE, DocCfg
+from epstein_files.documents.config.email_cfg import EmailCfg
 from epstein_files.util.helpers.string_helper import join_truthy, quote
 from epstein_files.people.names import *
 from epstein_files.util.constant.strings import *