+++ epstein_files/documents/config/categories/money.py
@@ -1,7 +1,8 @@
 """
 Epstein money related files. This category makes is_interesting = True.
 """
-from epstein_files.documents.config.doc_cfg import DocCfg, EmailCfg
+from epstein_files.documents.config.doc_cfg import DocCfg
+from epstein_files.documents.config.email_cfg import EmailCfg
 from epstein_files.people.names import *
 from epstein_files.util.constant.strings import *