+++ epstein_files/documents/config/categories/text_msg.py
@@ -1,4 +1,5 @@
-from epstein_files.documents.config.doc_cfg import CommunicationCfg, DocCfg
+from epstein_files.documents.config.communication_cfg import CommunicationCfg
+from epstein_files.documents.config.doc_cfg import DocCfg
 from epstein_files.people.names import *
 from epstein_files.util.constant.strings import *